+++ tests/functional/L0_Inplace_Optimize_torch_identity_multi_graph/test.py
@@ -49,8 +49,6 @@ def main():
     from tests import utils
     from tests.functional.common.utils import collect_optimize_statuses, validate_status
 
-    nav.inplace_config.mode = nav.Mode.RECORDING
-
     parser = argparse.ArgumentParser()
     parser.add_argument(
         "--status",
@@ -81,7 +79,7 @@ def main():
 
     model = Model()
     t = torch.randn(2, 3)
-    dataloader = [{"x": t}, {"x": t, "y": t}, {"x": t, "z": t}, {"x": t, "y": t, "z": t}] * 2
+    dataloader = [(1, {"x": t}), (1, {"x": t, "y": t}), (1, {"x": t, "z": t}), (1, {"x": t, "y": t, "z": t})] * 2
 
     def verify_func(ys_runner, ys_expected):
         for y_runner, y_expected in zip(ys_runner, ys_expected):
@@ -101,15 +99,9 @@ def main():
             "TensorRT",
         ),
     )
-    model = nav.Module(model, optimize_config=optimize_config, name="identity")
+    model = nav.Module(model, name="identity")
 
-    for batch in dataloader:
-        model(**batch)
-
-    nav.optimize()
-
-    for batch in dataloader:
-        model(**batch)
+    nav.optimize(func=model, dataloader=dataloader, config=optimize_config)
 
     names, packages = [], []
     for name, module in module_registry.items():